+++ weblog-rs/pages/components/document.rs
@@ -1,6 +1,6 @@
 use coyote::{attr_val, tmpl, Component};
 
-pub fn document_frame(
+pub fn document(
     language: Component,
     metas: Component,
     styles: Component,
@@ -24,21 +24,3 @@ pub fn document_frame(
     )
 }
 
-pub fn lang_en() -> Component {
-    attr_val("lang", "en-us")
-}
-
-pub fn metas(title: &str, description: &str) -> Component {
-    tmpl(
-        "
-        <meta charset=utf-8>
-        <meta name=viewport content=\"width=device-width, initial-scale=1\">
-        <meta {}>
-        <meta {}>
-        ",
-        [
-            attr_val("title", title),
-            attr_val("description", description),
-        ],
-    )
-}
