+++ weblog-rs/pages/components/type_flyweight.rs
@@ -1,0 +1,7 @@
+struct DocumentParams {
+    language: Component;
+    metas: Component;
+    styles: Component;
+    scripts: Component;
+    content: Component;
+}
